+++ saqc/core/translation/dmpscheme.py
@@ -16,6 +16,7 @@ import pandas as pd
 
 from saqc import BAD, DOUBTFUL, GOOD, UNFLAGGED
 from saqc.core import Flags, History
+from saqc.core.frame import DictOfSeries
 from saqc.core.translation.basescheme import BackwardMap, ForwardMap, MappingScheme
 from saqc.lib.tools import getUnionIndex
 
@@ -68,29 +69,29 @@ class DmpScheme(MappingScheme):
     def __init__(self):
         super().__init__(forward=self._FORWARD, backward=self._BACKWARD)
 
-    def toHistory(self, field_flags: pd.DataFrame):
+    def toHistory(self, flags: pd.DataFrame):
         """
         Translate a single field of external ``Flags`` to a ``History``
         """
-        field_history = History(field_flags.index)
+        history = History(flags.index)
 
-        for (flag, cause, comment), values in field_flags.groupby(_QUALITY_LABELS):
+        for (flag, cause, comment), values in flags.groupby(_QUALITY_LABELS):
             try:
                 comment = json.loads(comment)
             except json.decoder.JSONDecodeError:
                 comment = {"test": "unknown", "comment": ""}
 
-            histcol = pd.Series(np.nan, index=field_flags.index)
-            histcol.loc[values.index] = self(flag)
+            column = pd.Series(np.nan, index=flags.index)
+            column.loc[values.index] = self(flag)
 
             meta = {
                 "func": comment["test"],
                 "kwargs": {"comment": comment["comment"], "cause": cause},
             }
-            field_history.append(histcol, meta=meta)
-        return field_history
+            history.append(column, meta=meta)
+        return history
 
-    def toInternal(self, df: pd.DataFrame) -> Flags:
+    def toInternal(self, flags: pd.DataFrame | DictOfSeries) -> Flags:
         """
         Translate from 'external flags' to 'internal flags'
 
@@ -104,18 +105,23 @@ class DmpScheme(MappingScheme):
         Flags object
         """
 
-        self.validityCheck(df)
+        self.validityCheck(flags)
 
         data = {}
 
-        for field in df.columns.get_level_values(0).drop_duplicates():
-            data[str(field)] = self.toHistory(df[field])
+        if isinstance(flags, pd.DataFrame):
+            fields = flags.columns.get_level_values(0).drop_duplicates()
+        else:
+            fields = flags.columns
+
+        for field in fields:
+            data[str(field)] = self.toHistory(flags[field])
 
         return Flags(data)
 
     def toExternal(
         self, flags: Flags, attrs: dict | None = None, **kwargs
-    ) -> pd.DataFrame:
+    ) -> DictOfSeries:
         """
         Translate from 'internal flags' to 'external flags'
 
@@ -132,10 +138,7 @@ class DmpScheme(MappingScheme):
         """
         tflags = super().toExternal(flags, attrs=attrs)
 
-        out = pd.DataFrame(
-            index=getUnionIndex(tflags),
-            columns=pd.MultiIndex.from_product([flags.columns, _QUALITY_LABELS]),
-        )
+        out = DictOfSeries()
 
         for field in tflags.columns:
             df = pd.DataFrame(
@@ -163,13 +166,13 @@ class DmpScheme(MappingScheme):
                 df.loc[valid, "quality_comment"] = comment
                 df.loc[valid, "quality_cause"] = cause
 
-            out[field] = df.reindex(out.index)
+            out[field] = df
 
         self.validityCheck(out)
         return out
 
     @classmethod
-    def validityCheck(cls, df: pd.DataFrame) -> None:
+    def validityCheck(cls, flags: pd.DataFrame | DictOfSeries) -> None:
         """
         Check wether the given causes and comments are valid.
 
@@ -178,21 +181,16 @@ class DmpScheme(MappingScheme):
         df : external flags
         """
 
-        cols = df.columns
-        if not isinstance(cols, pd.MultiIndex):
-            raise TypeError("DMP-Flags need multi-index columns")
+        for df in flags.values():
 
-        if not cols.get_level_values(1).isin(_QUALITY_LABELS).all(axis=None):
-            raise TypeError(
-                f"DMP-Flags expect the labels {list(_QUALITY_LABELS)} in the secondary level"
-            )
+            if not df.columns.isin(_QUALITY_LABELS).all(axis=None):
+                raise TypeError(
+                    f"DMP-Flags expect the labels {list(_QUALITY_LABELS)} in the secondary level"
+                )
 
-        for field in df.columns.get_level_values(0):
-            # we might have NaN injected by DictOfSeries -> DataFrame conversions
-            field_df = df[field].dropna(how="all", axis="index")
-            flags = field_df["quality_flag"]
-            causes = field_df["quality_cause"]
-            comments = field_df["quality_comment"]
+            flags = df["quality_flag"]
+            causes = df["quality_cause"]
+            comments = df["quality_comment"]
 
             if not flags.isin(cls._FORWARD.keys()).all(axis=None):
                 raise ValueError(